Create README.md
This program works as a complete ordering system, where customers can make their food and beverage selections for breakfast, lunch or dinner, the order is then saved and sent to the kitchen section to start preparation, then once the order is prepared, it is sent to the waitstaff for billing to the customer. This program uses a mySQL database to store orders, with use of the JDBC API and the Java mySQL connector .jar file, therefore these are necessary for proper program execution.